HTML and CSS Projects
This is a display of sample websites that I made using HTML and CSS. 

Projects 
Space Station Site
Movie Theater Site

Space Station Site
This page is an example home page for a Space travel company. It's styled using CSS and contains a navbar, image containers, and a form. 

Movie Theater Site
This page is an example home page for a Move theater that is styled using Bootstrap 4. It contains many Bootstrap components including but not limited to a navbar, jumbotron, form, cards, and dropdowns.